+++ frontend/app/src/screens/object-item-details/action-buttons/index.tsx
@@ -2,6 +2,7 @@ import { StringParam, useQueryParam } from "use-query-params";
 import { DetailsButtons } from "./details-buttons";
 import { QSP } from "@/config/qsp";
 import { TASK_TAB } from "@/config/constants";
+import { RelationshipsButtons } from "./relationships-buttons";
 
 export function ActionButtons(props: any) {
   const [qspTab] = useQueryParam(QSP.TAB, StringParam);
@@ -11,7 +12,7 @@ export function ActionButtons(props: any) {
   }
 
   if (qspTab && qspTab !== TASK_TAB) {
-    return <>YES</>;
+    return <RelationshipsButtons {...props} />;
   }
 
   return null;
