+++ frontend/app/src/screens/object-item-details/action-buttons/relationships-buttons.tsx
@@ -1,0 +1,133 @@
+import { ButtonWithTooltip } from "@/components/buttons/button-primitive";
+import SlideOver, { SlideOverTitle } from "@/components/display/slide-over";
+import DynamicForm from "@/components/form/dynamic-form";
+import { SelectOption } from "@/components/inputs/select";
+import { Alert, ALERT_TYPES } from "@/components/ui/alert";
+import { QSP } from "@/config/qsp";
+import { ADD_RELATIONSHIP } from "@/graphql/mutations/relationships/addRelationship";
+import { usePermission } from "@/hooks/usePermission";
+import { genericsState, schemaState } from "@/state/atoms/schema.atom";
+import { useMutation } from "@apollo/client";
+import { Icon } from "@iconify-icon/react";
+import { useAtomValue } from "jotai";
+import { useState } from "react";
+import { useParams } from "react-router-dom";
+import { toast } from "react-toastify";
+import { StringParam, useQueryParam } from "use-query-params";
+
+export function RelationshipsButtons() {
+  const permission = usePermission();
+  const { objectKind, objectid } = useParams();
+  const [addRelationship] = useMutation(ADD_RELATIONSHIP);
+  const generics = useAtomValue(genericsState);
+  const schemaList = useAtomValue(schemaState);
+  const [relationshipTab] = useQueryParam(QSP.TAB, StringParam);
+
+  const parentSchema = schemaList.find((s) => s.kind === objectKind);
+  const parentGeneric = generics.find((s) => s.kind === objectKind);
+  const relationshipSchema = parentSchema?.relationships?.find((r) => r?.name === relationshipTab);
+  const relationshipGeneric = parentGeneric?.relationships?.find(
+    (r) => r?.name === relationshipTab
+  );
+  const relationshipSchemaData = relationshipSchema || relationshipGeneric;
+  const generic = generics.find((g) => g.kind === relationshipSchemaData?.kind);
+
+  const [showAddDrawer, setShowAddDrawer] = useState(false);
+
+  let options: SelectOption[] = [];
+
+  if (generic) {
+    (generic.used_by || []).forEach((kind) => {
+      const relatedSchema = schemaList.find((s) => s.kind === kind);
+
+      if (relatedSchema) {
+        options.push({
+          id: relatedSchema.kind,
+          name: relatedSchema.name,
+        });
+      }
+    });
+  } else {
+    const relatedSchema = schemaList.find((s) => s.kind === relationshipSchema?.peer);
+
+    if (relatedSchema) {
+      options.push({
+        id: relatedSchema.kind,
+        name: relatedSchema.label ?? relatedSchema.name,
+      });
+    }
+  }
+
+  const handleSubmit = async (data: any) => {
+    const { relation } = data;
+
+    if (relation?.id || relation?.from_pool) {
+      await addRelationship({
+        variables: {
+          objectId: objectid,
+          relationshipIds: [relation],
+          relationshipName: relationshipSchema?.name,
+        },
+      });
+
+      // TODO: REFETCH
+
+      toast(
+        <Alert
+          type={ALERT_TYPES.SUCCESS}
+          message={`Association with ${relationshipSchema?.peer} added`}
+        />
+      );
+
+      setShowAddDrawer(false);
+    }
+  };
+
+  return (
+    <>
+      <ButtonWithTooltip
+        disabled={!permission.write.allow}
+        tooltipEnabled
+        tooltipContent={permission.write.message ?? "Edit object"}
+        onClick={() => setShowAddDrawer(true)}
+        data-testid="edit-button">
+        <Icon icon="mdi:plus" className="mr-1.5" aria-hidden="true" /> Add{" "}
+        {relationshipSchema?.label ?? relationshipSchema?.kind ?? "relationship"}
+      </ButtonWithTooltip>
+
+      <SlideOver
+        title={
+          parentSchema && (
+            <SlideOverTitle
+              schema={parentSchema}
+              currentObjectLabel={relationshipSchema?.label}
+              title={`Associate a new ${relationshipSchema?.label}`}
+              subtitle={`Add a new ${relationshipSchema?.label} to the current object`}
+            />
+          )
+        }
+        open={showAddDrawer}
+        setOpen={setShowAddDrawer}>
+        <DynamicForm
+          fields={[
+            {
+              name: "relation",
+              label: relationshipSchema?.label,
+              type: "relationship",
+              relationship: { ...relationshipSchema, cardinality: "one", inherited: true },
+              schema: relationshipSchemaData,
+              options,
+            },
+          ]}
+          onSubmit={async ({ relation }) => {
+            await handleSubmit({ relation: relation.value });
+          }}
+          onCancel={() => {
+            setShowAddDrawer(false);
+          }}
+          className="w-full p-4"
+        />
+      </SlideOver>
+    </>
+  );
+}
